refactor: HomeController 생성 및 asis 코드 삭제
