+++ mvc/src/main/java/nextstep/mvc/controller/AnnotationHandlerAdapter.java
@@ -1,23 +1,20 @@
-package com.techcourse;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class ManualHandlerAdapter implements HandlerAdapter {
+public class AnnotationHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        return handler instanceof Controller;
+        return handler instanceof HandlerExecution;
     }
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
             throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
+        return ((HandlerExecution) handler).handle(request, response);
     }
 }
