--- mvc/src/main/java/nextstep/mvc/controller/tobe/AnnotationHandlerAdapter.java
@@ -1,20 +1,0 @@
-package nextstep.mvc.controller.tobe;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.view.ModelAndView;
-
-public class AnnotationHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof HandlerExecution;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
-            throws Exception {
-        return ((HandlerExecution) handler).handle(request, response);
-    }
-}
