+++ mvc/src/test/java/nextstep/mvc/controller/tobe/AnnotationHandlerAdapterTest.java
@@ -2,6 +2,8 @@ package nextstep.mvc.controller.tobe;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.mvc.controller.AnnotationHandlerAdapter;
+import nextstep.mvc.controller.HandlerExecution;
 import nextstep.mvc.controller.asis.Controller;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
